Create workingincommandline
I got the game to do everything I want it to in the command line.  Now I'll be trying to create the kivy app and getting it to work on that.  Any tips on how to make this code more efficient or changes that will translate better to kivy would be greatly appreciated.